New classes:
- QEdgeConstraint
- QExpand
- QToTPlanner
- QToQPlanner
- QToQSamplerPlanner

Removal of classes:

- StraightLinePathPlanner
- lazyprm

Renamed classes:

- PathPlanner is renamed to QToQPlanners for RRT.

The Ptr convention is used more extensively. Path planner interfaces
have all been changed.

macros.hpp has been restored.

CollisionDetector no longer takes ownership unless you ask for it
(using rw::common::ownedPtr of Ptr.hpp).



git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/RobWork/trunk@665 494c09c7-a67d-49d4-9938-151fb1965be8

+++ src/rw/loaders/rwxml/XMLRWLoader.cpp
@@ -584,9 +584,9 @@ namespace {
 
         // add all configurations, add home configs to default state
         BOOST_FOREACH(QConfig& config, dev._qconfig){
-            if(config.name=="Home"){
+            if(config.name == "Home") {
                 rw::math::Q q(config.q.size());
-                for(int i=0;i<q.size();i++)
+                for(size_t i = 0; i < q.size(); i++)
                     q[i] = config.q[i];
 
                 DeviceInitState *initDevState = new DeviceInitState(q,model);
